Works on issue 117.  This change is not really a speed performance but it simplifies the logic in eof_note_draw() by putting repeated code in a loop.  Consequently, this will make it easy to adapt this function to add a sixth fret.  I plan on changing the function skip the rendering and return either of the following values if the note (and its tail) are outside the viewable area: -1 : Note was clipped for being before the viewable area 1 : Note was clipped for being after the viewable area
The functions that render all notes in a loop can check the return value and if it's 1, it can stop rendering because no more notes in the chart would be visibly rendered.  This should significantly improve performance for rendering the piano roll, but the same thing should work for the 3D panel as well.