Works on issue 125.  The creation of the PART VOCALS track has been altered so that the order of phrasings is guaranteed to be:
Overdrive on, lyric phrase on, lyric, lyric pitch, ..., lyric phrase off, overdrive off

I had seen cases where it seemed that the lyric event would be written before the lyric phrase on, which could have caused FoFiX to read the phrase incorrectly.  I also applied logic to enforce a minimum amount of padding between a lyric phrase on event and the first lyric+pitch in the phrase, setting it for 5 deltas.  This is a macro, so it can be easily changed.